Cleaned up the way UseMod users are created.
UseMod basically used user IDs as session IDs.  Everytime an unknown person
(no cookie set) accessed the site, a new user ID would be locked for it,
and a cookie with that user ID would be set.

UseMod also reserved a bunch of randomish numbers for special purposes,
thus making the code even more unintelligible.  For example, an ID of 111
was a placeholder ID that signified that the cookie value was not set.
An ID of 113 showed that the cookie values did not correspond to the
recorded values.

One more lame thing that UseMod did: You needed to remember your user ID
in order to login.  The end result was that people would forget their
user IDs or passwords and would just create a new account.

createUsernameIndex.pl is a script that cleans up the wikidb/user
directory.  It eliminates locked IDs and duplicate accounts, and it
creates a BerkeleyDB database of usernames to IDs.

If a cookie is not set or if it doesn't correspond with information in
the database, then a User ID of 200 is used (which signifies guest user).
New IDs are only locked when people create accounts using the Preferences
button.

I also added a logout feature which deletes the session and clears the
cookie.