Sigh, this commit fixes a bug that's been there since I created the Database/*.pm classes.
I discovered the problem trying to get at old revisions of pages on
my wiki, the collab wiki, and purplewiki's wiki that had been spammed.

For pages where the number of the revision is not in sync with the 
count of revisions available no revision is retrieved. This is because
when I first did the migration, I only had new revisions, none had 
expired away. A look at my patch should explain more clearly what's 
going on. It's not the perfect fix, but it is a fix. I mention an
alternative in the FIXME.

In addition, the KeepDays setting in the default config is probably
too low for sites that are not busy. If it is at the default 14 days
it means that if someone comes along and spams a page of yours that is 
importnat but stale, you're rather screwed for pulling something out
of the KeptRevision, your only recourse is the diff_log.

I've set the default to 30 days. Wiki pages are small and disk is
cheap. It might make sense to make it even longer.