Mostly complete refactoring of the user system.  The code is just about ready for merging back into trunk, barring some cleanups here and there. Please take a look!
User database module is in PurpleWiki::Database::User::UseMod.pm.  At
some point, this can be abstracted into a base class so that others
can write replacement modules.  Added PurpleWiki::User.pm as the user
class.

Completed changing how the user system works.  UseMod created new users
with each session (essentially acting as session IDs), and reserved 
several UIDs for special purposes.  In my last set of changes, I changed
it to do real session handling (via CGI::Session) and removed most of the
special UIDs.  However, I reserved one -- 200 -- for guests.

Now that my changes are complete, there are no more reserved UIDs.
This is better from a customization and integration standpoint, because
you don't want to make assumption as to what UIDs the other system is
using, and you don't want to have a user system in parallel to the other
system if at all possible.  If a user is log in, the $user object will
be instantiated.  If no one has logged in, $user will be undef.

There's one known bug: Pages that are modified by guests (people who
are not logged in) do not show up when you view the history of the
page.

Other changes:

  * Added userdb.t test.

  * Changed the login functionality so that it uses usernames, not
    UIDs.  Also cleaned up the corresponding templates.

  * Fixed minor bug in parser that was causing phantom NIDs to appear
    in nonexistent pages.

  * Modified parser so that it doesn't require white space between
    the structural content and the {nid } tag.  This came out of an
    observation at Planetwork, where I noticed a lot of people not
    putting a space there.  The WikiText is normalized when you edit
    it again so that there is one space between the tag and the content.

  * Added ScriptDir in config, so that cookie is set for the correct
    path.  Before, the code assumed that it should read cookies for
    /cgi-bin/, which is not always the case.

  * Fixed wikihtml so that it doesn't show a warning.  Slightly modified
    the debug driver so that it displays NIDs for nodes that have them.