Added use warnings to all the tests. This just seems like good form. I'm not sure our current harness will even let us see the warnings, but I often run the tests by hand, so its good for that.
Added t/wikihtml.t (crap forgot manifest, will get in a moment)

wikihtml.t uses the same form as wikitext.t, and the same input.
I create output html file for some, but not all of the sampel files.
I got tree_test01 and tree_list which gets most of the nasties, and
all the hrs.

QUESTION: Does the moin moin style of <pre> marking do anything
inline?  If so, the tests are not currently good for that.

In the process I recognized some problems PurpleWiki::View::wikihtml
and cleaned them up.

First I merged in the changes from the user-refactor branch, revision
414. This puts line feeds in some places. I adjusted some of that,
adding a few more line feeds to make the output still cleaner.

Freelinks with line feeds were not being handled well. The linebreak
was disappearing in the id. So in addition to changing space to
underscore linefeed and return are too. This is to get the html
output/view to align with what appears to be the intention from the text
output/view.

Adjusted the handling of class attributes in links.

* made them all come after the href (at least that's what I was going for)
* made interwiki links have their own class (interwiki)
* gave freelinks their own class (freelink)

QUESTION: A colleague has requested that not yet created wiki pages
have a differnt class from those that do exist, so they can be css'ified
special. He finds the question mark hard to find or something.
He actually asked for the question mark to go away with the word 
having a special link. Good? Bad?

Further refinement of PurpleWiki::View::wikihtml.pm to get a variable
inside a loop.

I've not run all the tests yet after my changes. I'll commit another
cleanup in a moment.