This a rather sizable set of changes to support xhtml 1.1 compliance.
The biggest part of this is changing & to &amp; in URLs that show up
in HTML pages: including href attribute values.

The easiest way for me to do this was to change all the &sub calls
to sub() and then find the remain & and distinguish between those
that are part of regular expressions, those that are && and those
that are in html in some way. Getting rid of the &sub stuff was 
something I've been wanting to do anyway, so although I suspect
I could have done something that was a bit less destructive I 
figured there's no time like the present.

Fixed some tests to correspond with the new way.

Adjusted forms that have textareas with 'wrap="virtual"' to remove
the wrap attribute as xhtml does not allow it. This should have
effect in modern browsers.

Changed GET methods to get in html. XHTML is not a friend of the
upper case letter and invalidates it.

header.tt had two opening <html> tags.

remove a GetTextArea sub that wasn't being used.

added an alt tag containing the src href of an image to the img tag
as an alt is required

added a couple of dieS to Template/TT.pm to handle when a 
template file is not present.

tests are passing but there's probably still some ick. i'll 
change wiki.pl to be more generic (not include me info) when
I have it more polished.