This commit, regrettably, removes support for HTML::Template until such a time as we can work out a way to make the data structure being passed around for templating generic enough for HTML::Template's less code-in-templates style of doing things.
I think it is fairly close but it is not close enough for 
me to delay getting a 0.93 out and rather than leave 
some cruft around I'm just killing off the bits. We can always
pull them back out later if we want, or start from scratch--
it's not that much stuff.

To compensate I've made the calling of a template driver
configurable by adding TemplateDriver and TemplateDirectory config
options. If the driver is PurpleWiki::Template::TT then TemplateDriver
should be set to 'TT'.

This makes room for some future expansion &c