The tests that demonstrate the problem with a mixed list.
Plust a modified css file that incorporates some of the style
from planetwork plus some of the stuff some folk did in kb land
to make the wiki there a bit more palatable. I had planned to
make some changes to the templates to bring everything around
to a clean, simple, but not too terribly boring style, but 
I lost my oomph, so just style now. The font change alone
helps a lot. At default levels though, it's a bit large. but
setting it to something static would be against the religion, so...