http://purplewiki.blueoxen.net/forums/devel/2004-08/msg00038.html#nid017
This fixes the bug reported in the message linked above, it was
something like I suggested.

Basically $currentNid was scoped in an outer block for headers, so
headers could steal nids from one another when the next header did
not have a nid and the previous one did.

I removed currentNid from the outer scope, looked through all the
places where it is used to see that it is being somewhat smart.

And removed a few tabs.