upping the version to 0.94
fixing the bug wherein usernames are not listed in revision
histories, mentioned here:

 http://purplewiki.blueoxen.net/forums/devel/2004-08/msg00075.html

turns out someone replaced the folgers with something else on the
day i was working on Section.pm and I just plain forgot to
create subs for setting username or user id. Gets are there. So
now there are, and wiki.pl has been adjusted to use them.

I suspect there was some confusion on my part on when the info
comes from Page and when it doesn't. 

I hate the convoluted world of the UseMod database. Who knows, though,
maybe I made it worse? Or maybe the bug has been there since we started
from UseMod. In any case, it should be okay now.