Refactored session handling into PurpleWiki::Session.  The methods are stolen straight from CGI::Session, and in fact, simply call those methods by default.
Also readded CookieName to the config.  Chris had noted that having this
allowed you to support single sessions across multiple Wikis.

  http://purplewiki.blueoxen.net/forums/devel/2004-08/msg00044.html#nid02

This was not quite true at the time, but it is possible now by subclassing
Session (modifying it to use the same wikidb/sessions directory) and by
making sure CookieName is set to be the same across each Wiki.  More mods
are necessary to make this work right, unfortunately, because the visited
pages cache isn't multisite aware.

This was a necessary change to do the Wiki integration for GoLightly,
and Blue Oxen will need this feature for its collaboratories as well.

If CookieName is not set in config, wiki.pl will use SiteName instead.