The Database API should now be fully consistent with the Wiki page.  The tests have been updated and they all pass against the modified UseMod Database, as do the new runlog.pl tests.
There is also an early version of a Subversion based back end.  This backend
is intended to use the Repository and Filesystem binding to access and update
a local repository (i.e. you have to be able to access the repository directory
files and databases locally).

Syndication/Rss changed to move the view operation out of Pages (local from
getWikiHTML).

Parser/WikiText changed to fix some transclusion problems.  In retrospect,
the problem was probably fixed elsewhere, but I did fix some structural
problems.  $sequence and $url are no longer file scoped statics.

Transcusion is the same thing as Rss, move view code back here from the
Archive class.  In this case it was getPageNode that was replaced.

Database/Pages.pm was updated to make it consistent with the finalized API.

Search/Wiki took back the 'searchResult' method from Archive as a function.

runlog.pl  Updated to do a better job factoring date changes from the diffs.

page.t was fixed to match the Database API

wiki.pl no longer used $Now, and is refactored for the API.