convert.pl  -- new script to do a page database conversion.  It uses the API, so it should be general from/to any database class.  The basic algorithm is:
1.  $page->allPages gets a list of all Ids.
2.  $page->getRevisions($id) gets the revisions including meta-data
3.  Save all the revision information using the timestamp as the key.
4.  Process the saved information sorted by timestamp.  This preserves
    the ordering.
5.  $newpages->putPage(...) to write the revision to the new database.

This process does not preserve timestamps, but they should be in the same.
It wouldn't be that hard to preserve the timestamps, but the API doesn't
support it now.  For the conversion from UseMod to DefaultArchive, the
revision number should be preserved, but this is not guaranteed.

I've tested it with the database created by my r/runlog tests:

1. Generate a UseMod databse by running tests 0-58 with the Database/Pages
as the database module.  Config in r/config, database rDB.
2. Run the conversion with input using the config in 1. and the output using
new/config and newDB.
3. Run the tail of non-updating tests out/request.51-58.  These include
accessing the revision history of all four pages in the test data set.

This test passes with this version.