Merged changes in trunk from version 0.94 to current (rev 519) into the database-api-1 branch.  Only wiki.pl had to be resolved manually.  Most of the rest related to test configurations.
Perhaps I should add a few things about that.  I have split tests, test
test configuration and other data from the database directories.  This makes
it a lot easier to cleanup the database and start a new run (i.e. just
remove it all).  So the original unit tests in 't' use 'tDB' for the database
directory and the new 'runlog.pl' tests use 'r/config' for configuration
and 'rDB' for the database.  You can see from the configs that these are set
up to take some of the more fixed input files from wikidb to obviate the need
to make copies of them.

The UseModDatabase implementation works with the new API, but it isn't exactly
the same as before.  We eliminated the 'subpage' feature, although '/' should
be allowed in pageIds.

Thinking about it further, the conversion program is based on the new library
and API, so I don't think it will properly read a database with subpages,
unless by some miracle this doesn't require anything special.  This needs to
be addressed before we can really release this.