I was working on being able to read UseMod database 'subpages' when I discovered that there were some bugs in my code to eliminate simple date/time differences from the listings.  I was causing it to miss a lot of problems. In particular, the $archive->allPages wasn't right.  The old code was returning a list of hashes containing 'id' and 'pageName', and the new API interface returns just a sorted list of ids.  The diff problems was masking it, but it has now been fixed.
It seems this problem was also covering up some problems with recentChanges
and page history.  Most of this was around the hash keys of the items returned.
I've tried to make more of this consistent with the rest of the API (i.e. id
becomes pageId, summary becomes changeSummary and so on).

For the 'subpages' issue, if it sees an id with a '/' in it, and the file
exists before translating the '/' to '+', it will consider it valid and will
be able to read, list and update that page.  It will not create new pages that
way and will instead do the '/' to '+' translation to build the filename.

What this doesn't account for is if there are additional semantics around
subpages, or if there were page attributes linking the subpage and main page.
This functionality will need a bit more testing as I don't have any current
test cases for this.

PurpleWiki/Database.pm          Return just the ids for a page list.
                                Also, some recentChanges stuff
PurpleWiki/Syndication/Rss.pm   pageName -> pageId (recentChanges)
PurpleWiki/DefaultArchive.pm    Fixed allPages (was returning duplicate ids)
                                Also, recentChanges and getRevisions stuff
PurpleWiki/Database/Pages.pm    Handle subpages on read/list operations
                                And the recentChanges field names
PurpleWiki/Database/KeptRevision.pm  Same thing.
PurpleWiki/View/wikihtml.pm     FreeLinks weren't being properly translated
                                to pageIds for pageExists
wiki.pl                         Create the hashed for the template from the
                                list of ids
r/runlog.pl                     Fixed diff problems
r/runlogm.pl  
out/wiki* and templates         No longer use oldtime, use oldrev instead.

RSS function:  One thing I'm not completely sure of is what form of the pageId
is wanted for the <title> resource in the RSS output.  I was giving the
freelink with the words capitalized and spaces not underscore.  Now it gives
what I would call the pageId with the capitalization and underscores.  Easy
enough to fix if this isn't correct.