Added $page->deletePage($id) method to all three backends.
Added tests for delete to page*.t tests.
-> adds a second page
-> use both pageExists and allPages to test for presence of the new page
-> delete the second page.
-> use pageExists and allPages again to make sure it is gone.

Archive/PlainText.pm   Removes all pages and history.  Because recentChanges
actually looks at version timestamps the deleted page will be really gone.

Archive/UseMod.pm      Removes the page and keep files for the page.  Does not
update the recent change log, so this one will still show up in recentChanges
until it ages off the end.

Archive/Subversion.pm  Just deletes the name from the current revision, so the
history remains and it can be manually restored.  You should still be able to
fetch old versions if you reference the exact version.  Because recentChanges
is based on the repository logs the file would still be in the list, but the
most recent change would be the delete.  Current testing doesn't show how this
will display, but it shouldn't break anything.  Should probably be improved
to show that the change was a delete.