I decided to add a recentChanges test to page*.t after deleting a page.  The results are different for each as discussed in the last change summary.
I also discovered an inconsistency in the use of 'summary' vs. 'changeSummary'
in the recentChanges output.  Then I unwisely attempted to synchronize the
same output field from getRevisions, but I would have had to change the
templates to support it so I backed this part out.

In any case, everything seems to be working now.