Fixed some of the test data for the Subversion backend (I didn't get the revision change right and it created a conflict in the output where it shouldn't have been and I missed it).
Added code for updating the NID database.  This code isn't enabled with this
commit which will require a small change in wiki.pl.  It works as follows:

The API for this is added to PurpleWiki::Sequence:

$sequence->updateURL($url, $nids)  where $url is the new URL for @$nids

The NID list must be all of the NIDs that are associated with the URL, all
other NIDs that reference this URL will be removed from the database.

With this change, the NID index can be updated in two ways:
1) the old way, when you supply a URL to the parser, it passes it when it
create a new NID and getNext (Sequence.pm) adds the NID to the index, but
not the reverse index.
2) the new way, when you supply a URL to the putPage call, it will get the
list of NIDs from the new page and call $sequence->updateURL.  This reads
the reverse index, and removes any entries in the index for any NIDs in the
old set, but not in the new list.  Finally it updates the reverse index.

The old way is still enabled, so it behaves as before.  I added a couple of
tests to delete a couple of NIDs from one page, then add them to another,
expecting it not to work correctly.  After this change is committed, I'm going
to see if it works when I enable.  If it does, then we can roll this change
into the upcoming release, and I'll update the test results and wiki.pl, but if
it doesn't I can just not commit the small change to wiki.pl and it will remain
dormant untill it can be fixed.

PurpleWiki/Archive/*.pm    Added code to update the Sequence DB if URL supplied
PurpleWiki/Sequence.pm     As described above, added updateURL and reverse idx
r/runlogm.pl t/run*        Fixed a problem related to filtering revision numbers
t/out1/*                   Test data update for revision sequence problem
t/out*/*                   Added tests 75-81 to move some NIDS
t/conv.out/*               Forgot to add a couple conversion tests to SVN