Fixed bug where Recent Changes and View Revisions were not displaying the username of a page.  The reason is that the username is no longer stored with the page.  I modified wiki.pl to look up the username based on the userId stored with each page.  I also modified the relevant templates.
Note that this lookup -- and the way Recent Changes is implemented in the
PlainText backend in general -- will be a performance hog in certain
situations.  We're introducing a lot of these with this next release; in
fact, we've had a nice history of introducing performance hogs.  And that's
okay.  The major hog will disappear once PurpleWiki runs under mod_perl
(Gerry's new testing framework seems to show that we're mod_perl safe,
despite all of the globals we use), and we can optimize the rest away as
we get to it.