Created a generic PurpleWiki::Page class and modified the backend-specific Page classes (embedded in the PurpleWiki::Archive::* files) to subclass it.
Created a PurpleWiki::Archive::Base class and modified all the backends to
subclass it.

Fixed several bugs in the Archive classes uncovered by the above two 
changes as well as by adding 'use strict'.  I realize that subclassing
in Perl simply to define the interface isn't strictly necessary, but it's
good practice, and the bugs that this process uncovered shows why.  I
discovered lots of inconsistencies in how the different backends
implemented the API.

Updated MANIFEST.  Renamed $user to $userId in backendConvert.pl for
better clarity.