Fixed problems with recentChanges and getRevisions.  The problem relates to variable scope, and it seems to be a weirdness with perl.  Here is the situation:
sub foo {
  my $var = [];
  sub local_sub {
    ...
    push(@$var, ...);
  }

  ... get_logs(..., &local_sub);
}

In my understanding of perl 'my' variables, the reference inside local_sub
should always refer to the same variable as in foo, but it doesn't.  The
first time foo is run, $var gets bound in local_sub and the next time in,
a new $var is created in foo but in local_sub it still refers to the $var
from the first invocation of foo.

I moved the declaration of $var outside of foo and it works.  Earlier I had
made it a 'local' declaration inside foo, but when 'use strict' was turned on
it started complaining about needing a package for the local declaration.

In any case, it works now even if it is a bit odd.