First major milestone on the action-plugin branch.  Merged changes up to date in trunk (i.e. all the database-api-1 changes &c).  All tests pass except for runwikSVN which is Subversion based integration tests.  Should be a small issue, but I wanted to check this in sooner rather than later so I could begin to show and describe all of this to the PurpleWiki communit.
The object of this branch is to create a plugin fromwork for both services
and web requests.  I'll get into some details below, but this milestone is
supposed to just refactor and not add or change any functionality.  I am
proposing this for a foundation for the 0.96 changes.  With this foundation
it will be easier to persue a number of objectives relating to:
1) multiple page databases (mostly in terms of relationships between the
many wikis sothat you could search 'parent wikis' when do page lookups,
searches, etc., but handling multiple configurations in a single mod_perl
instance which isa lower priority),
2) changes to URL schema (I've done the parsing part of this, but it should
probably incorporate URL generation as well, so you change both together).
3) YAML based configuration (dovetails well with multiple, thus more complex
hierarchical configureations for 1).

Please jump in with critiques and feedback.  I'm not committed to any of it,
just the high level goals of creating a more modular and plugable architecture.
I'm happy to change it even in major ways to make it play with existing
related projects as well as supporting other future stuff I haven't thought
about.

Refactor Summary:

wiki.pl has been very much simplified.  It still has a couple of supporting
routines, but most of it has been moved out either to the new 'action plugins'
in PurpleWiki/Action or the request code (PurpleWiki/ParseRequest.pm and
PurpleWiki/Request.pm).

There are two sets of plugins ($config->Driver and $config->Action) that are
hashes keyed by a module identifier and value of the name of a perl module
to load.  The Driver modules are loaded first, and after loading each, a
'service object' is created with the module's 'new' method (the config is
always passed as an argument to new) and that object is stored in a 'context'
hash keyed by the module identifier.

The Action plugins are loaded, and a 'register' function is called to add
entries to the 'action' handlers (they are registered with the request
parser).

To handle a request, the request parser is called ($context{request}) either
with CGI and session variables or with command line args.  The parser
returns a new PurpleWiki::Request object that is the single argument passed
to the action plugins.  $request->action() will return a code reference to
a handler for the request (i.e. the handler registered for the action the
parser puts in the request object).

Oh, and I change the package in wiki.pl to "PurpleWiki".  The actions are
put in this namespace as well so they can call common routines without
having to put in the package name.

In view, transclusion and other code where I was getting the 'archive'
instance from $config->{pages} as a temporary measure, I now pass the archive
object to the $tree->view() instances where it is needed.  I think this can
be improved by using the service-object / context-hash framework.