Added some skeleton docs for PurpleWiki::Archive::Sequence.
Refactored the command-line options for backendConvert.pl, replacing
the homegrown parsing with Getopt::Std.  Started initial work of 
integrating sequenceIndex.pl functionality into backendConvert.pl.
One big change remains.  sequenceIndex.pl rebuilds the index with no
thought of the other apps that use the sequence database.  It needs
to rebuild only the Wiki parts of the index, retaining the rest of the
index as well as the last used NID.